update2
update2

+++ Pb-1/#1.java
@@ -1,23 +1,23 @@
 //#1
 
   public class HelloWorld{
-  public static void main(String []args){
+    public static void main(String []args){
         
-  System.out.println(summ(19564));
+      System.out.println(summ(19564));
          
          
      }
-  private static int summ(int n){
-  int num =0;
+    private static long summ(long n){
+      long num =0;
         
-  for(int i=3;i < n;i++){
-    if(i % 3 == 0 || i % 5 == 0){
-        num+=i;
+        for(long i=3;i < n;i++){
+          if(i % 3 == 0 || i % 5 == 0){
+            num+=i;
                 
-        }
+           }
             
-    }
-      return num;
+         }
+           return num;
         
- }
-} 
+     }
+   }
